refactor(pages): separate logic from UI and reorganize educator/advisory components
- Moved logic to custom hooks: AdvisoryProgressLogic.js, EducatorPortalLogic.js, ReportIssue.js
- Renamed EducatorPortal.jsx to shared EducationPortal.jsx for both educator and advisory board
- Refactored ReportIssues.jsx and ViewProgressBoard to use new logic files
- Updated routes and components to use new logic structure
- Cleaned up AdvisorPortal.jsx and Portal.jsx to reduce role-based duplication
- Improved component reusability and scalability for future role-based expansion

known-bug: toast displays on login page after logout when no form action has occurred
