+++ frontend/src/App.jsx
@@ -1,6 +1,7 @@
 import './App.css';
 import './index.css'
 import AppRoutes from './Routes.jsx';
+import {ToastContainer} from 'react-toastify'
 import NavBar from './components/NavBar.jsx';
 import { useLocation } from 'react-router-dom';
 import PortalLogoBar from './components/PortalLogoBar.jsx';
@@ -24,15 +25,19 @@ function App() {
     // Set the role for use in specific portal returns based on role
     const [role, setRole] = useState('no-role-set');
 
+    // Prop for errors when the user does not select a valid 
+    const [roleError, setRoleError] = useState('');
+
     const isValid = () => {
-        return noNavigation.includes(location.pathname) == false && validRoutes.includes(location.pathname) == true;
+        return noNavigation.includes(location.pathname) === false && validRoutes.includes(location.pathname) === true;
     }
 
     return (
         <div className='flex'>
             {isValid() && <NavBar role={role}/>}
             {isValid() && <PortalLogoBar/>}
-            <AppRoutes role={role} setRole={setRole}/>
+            <AppRoutes role={role} setRole={setRole} roleError={roleError} setRoleError={setRoleError}/>
+            <ToastContainer autoClose={4000} stacked /> {/* Added logic to show toast success/error messages */}
         </div>
     );
 }
