--- frontend/src/pages/Advisory/AdvisoryProgress.jsx
@@ -1,103 +1,0 @@
-import React, { useState } from 'react';
-import { PieChart, Pie, Cell, Tooltip, Legend, ResponsiveContainer } from 'recharts';
-import '../../styles/AdvisoryModerator/pages/AdvisoryProgress.css';
-
-const dummyOverallData = [
-  { name: 'Completed', value: 50 },
-  { name: 'In Progress', value: 30 },
-  { name: 'Not Started', value: 20 },
-];
-
-const dummySessionBreakdownData = [
-  { name: 'Session A', value: 80 },
-  { name: 'Session B', value: 60 },
-  { name: 'Session C', value: 40 },
-];
-
-const COLORS = ['#0088FE', '#00C49F', '#FFBB28'];
-
-const ViewProgressBoard = ({ role }) => {
-  const [showSpreadsheetDropdown, setShowSpreadsheetDropdown] = useState(false);
-
-  const handleSpreadsheetClick = () => {
-    setShowSpreadsheetDropdown(!showSpreadsheetDropdown);
-  };
-
-  const hasStudents = true;
-  const hasSessionData = true;
-
-  if (!hasStudents) {
-    return (
-      <div className="view-progress-board">
-        <h2 className="error-message">No students in the classroom yet.</h2>
-      </div>
-    );
-  }
-
-  if (!hasSessionData) {
-    return (
-      <div className="view-progress-board">
-        <h2 className="error-message">No session data available yet.</h2>
-      </div>
-    );
-  }
-
-  return (
-    <div className="outer-container">
-      <div className="white-container">
-        <div className="view-progress-board">
-          <div className="header">
-            <h2 className="title">View Progress</h2>
-            <div className="spreadsheet-container">
-              <button className="spreadsheet-button" onClick={handleSpreadsheetClick}>
-                Spreadsheet
-              </button>
-              {showSpreadsheetDropdown && (
-                <div className="spreadsheet-dropdown">
-                  <div className="dropdown-item" onClick={() => console.log("Download Entire Class Spreadsheet")}>
-                    Entire Class
-                  </div>
-                  <div className="dropdown-item" onClick={() => console.log("Download Selected Student Spreadsheet")}>
-                    Selected Student
-                  </div>
-                </div>
-              )}
-            </div>
-          </div>
-          <div className="charts-container">
-            <div className="chart-card">
-              <h3 className="chart-title">Overall Performance</h3>
-              <ResponsiveContainer width="100%" height={300}>
-                <PieChart>
-                  <Pie data={dummyOverallData} cx="50%" cy="50%" outerRadius={80} dataKey="value" label>
-                    {dummyOverallData.map((entry, index) => (
-                      <Cell key={`cell-${index}`} fill={COLORS[index % COLORS.length]} />
-                    ))}
-                  </Pie>
-                  <Tooltip />
-                  <Legend />
-                </PieChart>
-              </ResponsiveContainer>
-            </div>
-            <div className="chart-card">
-              <h3 className="chart-title">Session Breakdown</h3>
-              <ResponsiveContainer width="100%" height={300}>
-                <PieChart>
-                  <Pie data={dummySessionBreakdownData} cx="50%" cy="50%" outerRadius={80} dataKey="value" label>
-                    {dummySessionBreakdownData.map((entry, index) => (
-                      <Cell key={`cell-session-${index}`} fill={COLORS[index % COLORS.length]} />
-                    ))}
-                  </Pie>
-                  <Tooltip />
-                  <Legend />
-                </PieChart>
-              </ResponsiveContainer>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default ViewProgressBoard;
