+++ frontend/src/pages/Portal.jsx
@@ -1,26 +1,35 @@
-import { useLocation } from 'react-router-dom';
+import { useLocation, Navigate } from 'react-router-dom';
+import {useEffect} from "react";
 import StudentPortal from './Student/StudentPortal.jsx';
-import EducatorPortal from './Educator/Portal';
+import EducationPortal from './AdvisoryEducator/EducationPortal.jsx';
 import GuardianPortal from "./Guardian/GuardianPortal.jsx";
 import ModeratorPortal from './Moderator/ModeratorPortal.jsx';
-import AdvisorPortal from './Advisory/AdvisorPortal.jsx';
 
-const Portal = () => {
-    // 1. Grab the useNavigate() data passed into /portal route with useLocation()
-    const location = useLocation();
-    const role = location.state?.role;
+// eslint-disable-next-line react/prop-types
+const Portal = ({ setRoleError }) => {
+  const location = useLocation();
+  const role = location.state?.role;
 
-    // 2. Map the role with the correct portal
-    const dict = {
-        "student": <StudentPortal/>,
-        "educator": <EducatorPortal/>,
-        "guardian": <GuardianPortal/>,
-        "moderator": <ModeratorPortal/>,
-        "advisor": <AdvisorPortal/>
-    };
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  const dict = {
+    "student": <StudentPortal />,
+    "educator": <EducationPortal />,
+    "guardian": <GuardianPortal />,
+    "moderator": <ModeratorPortal />,
+    "advisor": <EducationPortal />
+  };
 
-    // 3. Return the proper portal
-    return dict[role] || <div>invalid role</div>;
-}
+  useEffect(() => {
+    if (!(role in dict)) {
+      setRoleError("Account type not selected.");
+    }
+  }, [role, dict, setRoleError]);
+
+  if (role in dict) {
+    return dict[role];
+  } else {
+    return <Navigate to="/" />;
+  }
+};
 
 export default Portal;
